Add files via upload
Uploading project proposal that we all worked on and contributed to. If you want to see individual contributions please request from Zach as he is the owner of this document on google documents and he can grant you permissions so you can see what everyone worked on and when. It has the revision history.